# InvoSmart/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "InvoSmart, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Authentication/LoginClient": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "clientEmail",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "clientPassword",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Authentication/SignupClient": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "newClientFName",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "newClientLName",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "newClientCompanyName",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "newClientCompanyAddress",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "newClientCompanyCity",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "newClientCompanyState",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "newClientCompanyZipCode",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "newClientPhoneNumber",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "newClientEmail",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "newClientPassword",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Invoice/CreateInvoice": {
      "post": {
        "tags": [
          "Invoice"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/InvoiceData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetWeatherForecast",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/WeatherForecast"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "InvoiceData": {
      "type": "object",
      "properties": {
        "clientID": {
          "format": "int32",
          "type": "integer"
        },
        "invoiceNumber": {
          "type": "string"
        },
        "invoiceDate": {
          "format": "date-time",
          "type": "string"
        },
        "dueDate": {
          "format": "date-time",
          "type": "string"
        },
        "billedToEntityName": {
          "type": "string"
        },
        "billedToEntityAddress": {
          "type": "string"
        },
        "payableTo": {
          "type": "string"
        },
        "servicesRendered": {
          "type": "string"
        },
        "submittedOn": {
          "format": "date-time",
          "type": "string"
        },
        "subTotal": {
          "format": "double",
          "type": "number"
        },
        "total": {
          "format": "double",
          "type": "number"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/InvoiceItem"
          }
        }
      }
    },
    "InvoiceItem": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "qty": {
          "format": "int32",
          "type": "integer"
        },
        "unitPrice": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "WeatherForecast": {
      "type": "object",
      "properties": {
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "temperatureC": {
          "format": "int32",
          "type": "integer"
        },
        "temperatureF": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "summary": {
          "type": "string"
        }
      }
    }
  }
}